requirements.txt
absl-py==1.0.0

asgiref==3.5.0

astunparse==1.6.3

blessings==1.7

branca==0.5.0
Brotli==1.0.9
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8

click==8.1.2

dash==2.3.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
earthengine-api==0.1.306
ee==0.2
Flask==2.1.1
Flask-Compress==1.11
flatbuffers==2.0
folium==0.12.1.post1
fonttools==4.25.0
future==0.18.2
gast==0.5.3
gcloud==0.18.3
google-api-core==2.7.2
google-api-python-client==1.12.11
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
gunicorn==20.1.0
h5py==3.6.0
httplib2==0.20.4
httplib2shim==0.0.3

importlib==1.0.4
importlib-metadata==4.11.3
itsdangerous==2.1.2

keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.3.6
mkl-fft==1.3.1
mkl-service==2.4.0
munkres==1.1.4
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
patsy==0.5.2
pep517==0.12.0
Pillow==8.4.0
pip-tools==6.6.0
plotly==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.3
pywin32==302
requests-oauthlib==1.3.1
rsa==4.8

sip==4.19.13

sqlparse==0.4.2

statsmodels==0.13.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0

tf-estimator-nightly==2.8.0.dev2021122109

tomli==2.0.1

tzdata==2022.1
uritemplate==3.0.1
urllib3==1.26.7

webencodings==0.5.1
Werkzeug==2.1.1

wincertstore==0.2
windows-curses==2.3.0
wrapt==1.14.0
zipp==3.8.0
